requirements.txt
openpyxl
pymongo
pandas
json
python-docx